renamed adjacency_power_generators.py to adjacency_generators.py
