+++ stellargraph/mapper/__init__.py
@@ -28,3 +28,4 @@ from .sampled_link_generators import *
 from .sampled_node_generators import *
 from .full_batch_generators import *
 from .mini_batch_node_generators import *
+from .adjacency_generators import *
