+++ stellargraph/mapper/adjacency_generators.py
